Implement eContainingFeature (#7)
This commit overloads the elementID message field to allow for a new
syntax ('uri#fragment'), where 'uri' is a metamodel URI in the package
registry, and 'fragment'  identifies an element within that package.

This is mostly to allow for simple implementation of
'.eContainingFeature', '.eClass' and so on, where we need to fetch
objects from the metamodel. Regular MDObjects will still use the
MagicDraw UUIDs.